requirements-dev.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
black==23.10.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dataclasses-json==0.6.1
distro==1.8.0
fastapi==0.104.0
flake8==6.1.0
frozenlist==1.4.0
grpcio==1.59.0
grpcio-tools==1.59.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.333
langsmith==0.0.62
marshmallow==3.20.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
portalocker==2.8.2
protobuf==4.24.4
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
Pygments==2.16.1
pytest==7.4.3
pytest-subtests==0.11.0
python-dotenv==1.0.0
PyYAML==6.0.1
qdrant-client==1.6.3
regex==2023.10.3
requests==2.31.0
ruff==0.1.3
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.26.18
uvicorn==0.23.2
yarl==1.9.2
